requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
requests>=2.26.0
streamlit>=1.24.0 